requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
exceptiongroup==1.1.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.1
packaging==23.1
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg==3.1.9
PyJWT==2.7.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
